Struct google_cloudresourcemanager2::api::FolderMethods

pub struct FolderMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on folder resources. It is not used directly, but through the CloudResourceManager hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_cloudresourcemanager2 as cloudresourcemanager2;
 
use std::default::Default;
use oauth2;
use cloudresourcemanager2::CloudResourceManager;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = CloudResourceManager::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `get(...)`, `get_iam_policy(...)`, `list(...)`, `move_(...)`, `patch(...)`, `search(...)`, `set_iam_policy(...)`, `test_iam_permissions(...)` and `undelete(...)`
// to build up your call.
let rb = hub.folders();

Implementations

impl<'a, C> FolderMethods<'a, C>

pub fn create(&self, request: Folder) -> FolderCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a Folder in the resource hierarchy. Returns an Operation which can be used to track the progress of the folder creation workflow. Upon success the Operation.response field will be populated with the created Folder. In order to succeed, the addition of this new Folder must not violate the Folder naming, height or fanout constraints. + The Folder's display_name must be distinct from all other Folders that share its parent. + The addition of the Folder must not cause the active Folder hierarchy to exceed a height of 10. Note, the full active + deleted Folder hierarchy is allowed to reach a height of 20; this provides additional headroom when moving folders that contain deleted folders. + The addition of the Folder must not cause the total number of Folders under its parent to exceed 300. If the operation fails due to a folder constraint violation, some errors may be returned by the CreateFolder request, with status code FAILED_PRECONDITION and an error description. Other folder constraint violations will be communicated in the Operation, with the specific PreconditionFailure returned via the details list in the Operation.error field. The caller must have resourcemanager.folders.create permission on the identified parent.

Arguments

request - No description provided.

pub fn delete(&self, name: &str) -> FolderDeleteCall<'a, C>

Create a builder to help you perform the following task:

Requests deletion of a Folder. The Folder is moved into the DELETE_REQUESTED state immediately, and is deleted approximately 30 days later. This method may only be called on an empty Folder in the ACTIVE state, where a Folder is empty if it doesn't contain any Folders or Projects in the ACTIVE state. The caller must have resourcemanager.folders.delete permission on the identified folder.

Arguments

name - Required. the resource name of the Folder to be deleted. Must be of the form folders/{folder_id}.

pub fn get(&self, name: &str) -> FolderGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves a Folder identified by the supplied resource name. Valid Folder resource names have the format folders/{folder_id} (for example, folders/1234). The caller must have resourcemanager.folders.get permission on the identified folder.

Arguments

name - Required. The resource name of the Folder to retrieve. Must be of the form folders/{folder_id}.

pub fn get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> FolderGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a Folder. The returned policy may be empty if no such policy or resource exists. The resource field should be the Folder's resource name, e.g. "folders/1234". The caller must have resourcemanager.folders.getIamPolicy permission on the identified folder.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn list(&self) -> FolderListCall<'a, C>

Create a builder to help you perform the following task:

Lists the Folders that are direct descendants of supplied parent resource. List provides a strongly consistent view of the Folders underneath the specified parent resource. List returns Folders sorted based upon the (ascending) lexical ordering of their display_name. The caller must have resourcemanager.folders.list permission on the identified parent.

pub fn move_(
&self,
request: MoveFolderRequest,
name: &str
) -> FolderMoveCall<'a, C>

Create a builder to help you perform the following task:

Moves a Folder under a new resource parent. Returns an Operation which can be used to track the progress of the folder move workflow. Upon success the Operation.response field will be populated with the moved Folder. Upon failure, a FolderOperationError categorizing the failure cause will be returned - if the failure occurs synchronously then the FolderOperationError will be returned via the Status.details field and if it occurs asynchronously then the FolderOperation will be returned via the Operation.error field. In addition, the Operation.metadata field will be populated with a FolderOperation message as an aid to stateless clients. Folder moves will be rejected if they violate either the naming, height or fanout constraints described in the CreateFolder documentation. The caller must have resourcemanager.folders.move permission on the folder's current and proposed new parent.

Arguments

request - No description provided.

name - Required. The resource name of the Folder to move. Must be of the form folders/{folder_id}

pub fn patch(&self, request: Folder, name: &str) -> FolderPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a Folder, changing its display_name. Changes to the folder display_name will be rejected if they violate either the display_name formatting rules or naming constraints described in the CreateFolder documentation. The Folder's display name must start and end with a letter or digit, may contain letters, digits, spaces, hyphens and underscores and can be between 3 and 30 characters. This is captured by the regular expression: \p{L}\p{N}{1,28}[\p{L}\p{N}]. The caller must have resourcemanager.folders.update permission on the identified folder. If the update fails due to the unique name constraint then a PreconditionFailure explaining this violation will be returned in the Status.details field.

Arguments

request - No description provided.

name - Output only. The resource name of the Folder. Its format is folders/{folder_id}, for example: "folders/1234".

pub fn search(&self, request: SearchFoldersRequest) -> FolderSearchCall<'a, C>

Create a builder to help you perform the following task:

Search for folders that match specific filter criteria. Search provides an eventually consistent view of the folders a user has access to which meet the specified filter criteria. This will only return folders on which the caller has the permission resourcemanager.folders.get.

Arguments

request - No description provided.

pub fn set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> FolderSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on a Folder, replacing any existing policy. The resource field should be the Folder's resource name, e.g. "folders/1234". The caller must have resourcemanager.folders.setIamPolicy permission on the identified folder.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> FolderTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified Folder. The resource field should be the Folder's resource name, e.g. "folders/1234". There are no permissions required for making this API call.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn undelete(
&self,
request: UndeleteFolderRequest,
name: &str
) -> FolderUndeleteCall<'a, C>

Create a builder to help you perform the following task:

Cancels the deletion request for a Folder. This method may only be called on a Folder in the DELETE_REQUESTED state. In order to succeed, the Folder's parent must be in the ACTIVE state. In addition, reintroducing the folder into the tree must not violate folder naming, height and fanout constraints described in the CreateFolder documentation. The caller must have resourcemanager.folders.undelete permission on the identified folder.

Arguments

request - No description provided.

name - Required. The resource name of the Folder to undelete. Must be of the form folders/{folder_id}.